+++ user/src/main/java/com/tradingtrends/user/presentation/request/DeleteRequestDto.java
@@ -1,0 +1,10 @@
+package com.tradingtrends.user.presentation.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class DeleteRequestDto {
+    private String password;
+}
